{"schema":"decodex/commit/1","summary":"Extract macOS native shell window owner","authority":"manual"}
