+++ docs/reference/workspace-layout.md
@@ -115,6 +115,9 @@ Key paths:
   renderer phase timing telemetry lives under `overlay/rendering/timing.rs`
 - `packages/rsnap-overlay/src/overlay/rendering/affordances/toolbar.rs`: Frozen toolbar rendering
   entry; geometry and tool-palette policy live under `overlay/rendering/affordances/toolbar/`
+- `packages/rsnap-overlay/src/overlay/macos_native_capture_shell_runtime.rs`: retained macOS
+  native capture-shell host sync; AppKit shell window lifecycle and creation live under
+  `overlay/macos_native_capture_shell_runtime/shell_windows.rs`
 - `packages/rsnap-overlay/src/live_frame_stream_macos.rs`: current macOS live-stream support
 - `packages/rsnap-overlay/src/scroll_capture.rs`: current scroll-capture session entry with
   focused support modules under `scroll_capture/`
